+++ programs/lazy-distributor/src/instructions/initialize_lazy_distributor_v0.rs
@@ -5,10 +5,10 @@ use anchor_spl::{
 };
 use circuit_breaker::{
   cpi::{accounts::InitializeAccountWindowedBreakerV0, initialize_account_windowed_breaker_v0},
-  CircuitBreaker, InitializeAccountWindowedBreakerArgsV0,
+  CircuitBreaker, InitializeAccountWindowedBreakerArgsV0, WindowedCircuitBreakerConfigV0,
 };
 
-use crate::{circuit_breaker::WindowedCircuitBreakerConfigV0, state::*};
+use crate::state::*;
 
 #[derive(AnchorSerialize, AnchorDeserialize, Clone, Default)]
 pub struct InitializeLazyDistributorArgsV0 {
@@ -78,7 +78,7 @@ pub fn handler(
     ),
     InitializeAccountWindowedBreakerArgsV0 {
       authority: args.authority,
-      config: args.window_config.into(),
+      config: args.window_config,
       owner: ctx.accounts.lazy_distributor.key(),
     },
   )?;