+++ CHANGELOG.md
@@ -3,6 +3,78 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.10.13](https://github.com/helium/helium-program-library/compare/v0.10.12...v0.10.13) (2025-05-12)
+
+**Note:** Version bump only for package helium-program-library
+
+
+
+
+
+## [0.10.12](https://github.com/helium/helium-program-library/compare/v0.10.11...v0.10.12) (2025-05-05)
+
+**Note:** Version bump only for package helium-program-library
+
+
+
+
+
+## [0.10.11](https://github.com/helium/helium-program-library/compare/v0.10.10...v0.10.11) (2025-04-30)
+
+**Note:** Version bump only for package helium-program-library
+
+
+
+
+
+## [0.10.10](https://github.com/helium/helium-program-library/compare/v0.10.9...v0.10.10) (2025-04-30)
+
+**Note:** Version bump only for package helium-program-library
+
+
+
+
+
+## [0.10.9](https://github.com/helium/helium-program-library/compare/v0.10.8...v0.10.9) (2025-04-30)
+
+**Note:** Version bump only for package helium-program-library
+
+
+
+
+
+## [0.10.8](https://github.com/helium/helium-program-library/compare/v0.9.34...v0.10.8) (2025-04-29)
+
+
+
+## [0.10.3](https://github.com/helium/helium-program-library/compare/v0.10.2...v0.10.3) (2025-03-25)
+
+
+
+## [0.10.2](https://github.com/helium/helium-program-library/compare/v0.10.1...v0.10.2) (2025-03-24)
+
+
+
+## [0.10.1](https://github.com/helium/helium-program-library/compare/v0.10.0...v0.10.1) (2025-03-21)
+
+
+
+# [0.10.0](https://github.com/helium/helium-program-library/compare/v0.9.31...v0.10.0) (2025-03-20)
+
+**Note:** Version bump only for package helium-program-library
+
+
+
+
+
+## [0.10.7](https://github.com/helium/helium-program-library/compare/v0.10.6...v0.10.7) (2025-04-22)
+
+**Note:** Version bump only for package helium-program-library
+
+
+
+
+
 ## [0.10.6](https://github.com/helium/helium-program-library/compare/v0.10.5...v0.10.6) (2025-04-08)
 
 **Note:** Version bump only for package helium-program-library